requirements.txt
apturl==0.5.2
asttokens==2.2.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
executing==1.2.0
fasteners==0.14.1
Flask==2.3.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==3.12.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.2
reportlab==3.6.8
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
stack-data==0.6.2
systemd-python==234
tornado==6.3.1
traitlets==5.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.6
Werkzeug==2.3.4
xdg==5
xkit==0.0.0
zipp==1.0.0
